rsync: Initial revision
Add an initial version of the rsync engine. This engine will
synchronize directories between a local directory and a remote rsync
server; it can operate using either SSH or the normal rsync protocol.

This engine properly handles added and deleted files, but not modified
files; it is thus only suitable for archival purposes.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>